requirements.txt
streamlit
groq
requests
python-dotenv
